google: dedupe calendar/gmail api_message and mark py binding twin (#3911)
Move the Google REST error-envelope extractor (api_message plus the
ApiErrorBody/ApiErrorDetail wire types and its unit tests) into
google-auth, the crate both clients already share, and point calendar
and gmail's check_status at it. That clears the T1 clone
(calendar/src/lib.rs:264-276 = gmail/src/lib.rs:259-271).

The py T2 twin (get_message/get_thread wrappers) was already merged
into one parameterized helper (gmail_resource_future, #2618); what the
scanner still sees is the per-method pyo3 declaration, which cannot
merge without changing the Python API. Mark it clone:ignore with that
justification, following the minecraft protocol bindings precedent.

Repo scan: duplicated_lines 2216 -> 2189.

Closes #3911.

(authored by an AI agent via Claude Code, model fable)
